Moved windowUtils into separate files.
